Added custom prior function similar to prior_from_file from sherpa
I needed to use the posteriors from another bxa fit as a prior for a new fit. Sherpa has a prior_from_file function which allows you to do this, but for some reason it was missing as a function in the PyXspec implementation. I tried to keep the code as familiar as possible to the sherpa version.